+++ commercetools/commercetools-sdk-java-api/src/test/java-generated/com/commercetools/api/models/project/SearchIndexingConfigurationTest.java
@@ -31,7 +31,10 @@ public class SearchIndexingConfigurationTest {
                         .productsSearch(
                             new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl()) },
                 new Object[] { SearchIndexingConfiguration.builder()
-                        .orders(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl()) } };
+                        .orders(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl()) },
+                new Object[] { SearchIndexingConfiguration.builder()
+                        .customers(
+                            new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl()) } };
     }
 
     @Test
@@ -55,6 +58,14 @@ public class SearchIndexingConfigurationTest {
         SearchIndexingConfiguration value = SearchIndexingConfiguration.of();
         value.setOrders(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl());
         Assertions.assertThat(value.getOrders())
+                .isEqualTo(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl());
+    }
+
+    @Test
+    public void customers() {
+        SearchIndexingConfiguration value = SearchIndexingConfiguration.of();
+        value.setCustomers(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl());
+        Assertions.assertThat(value.getCustomers())
                 .isEqualTo(new com.commercetools.api.models.project.SearchIndexingConfigurationValuesImpl());
     }
 }